Refactor file names and locations
