+++ src/utils/is-empty.ts
